Updated and reprogrammed almost all pages
I worked for several months reprogramming and improving the project overall. It is now more aesthetically pleasing and functional in every way.
New features include book search, book reading, personal profile, and homepage.
It is still in development, with only a few features missing, such as adding your own book to the website and implementing a "continue reading" feature.